First Upload
1. API integration complete with search inputs and movie description
load all working well.
2. While not a requirement to this project, I continue to have issues
with the pagination build and can’t seem to get the scroll
functionality quite down yet as it at times will go over and then under
the allowed amount.
4. I may decide to branch out here and make a simple “left arrow”
“right arrow” integration to create a less complicated build however
will try and complete the current plan as is with an easy to use
pagination scroll and click feature that loads new requests each time.